bottom bar and initial page to check the user added!
