+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,13 @@
         android:theme="@style/Theme.ElearningMAD"
         tools:targetApi="31">
         <activity
+            android:name=".ui.EditProfile"
+            android:exported="false"
+            android:theme="@style/Theme.AppCompat.DayNight.NoActionBar"/>
+        <activity
+            android:name=".ui.InitialPage"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name">
@@ -21,6 +28,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
