+++ app/src/main/java/com/example/elearningmad/MainActivity.kt
@@ -1,16 +1,13 @@
 package com.example.elearningmad
 
+import android.content.Intent
 import android.os.Bundle
-import android.view.Window
-import android.view.WindowManager
-import androidx.appcompat.app.ActionBar
-import com.google.android.material.bottomnavigation.BottomNavigationView
+
+import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
 import com.example.elearningmad.databinding.ActivityMainBinding
+import com.example.elearningmad.ui.InitialPage
+
 
 class MainActivity : AppCompatActivity() {
 
@@ -23,17 +20,27 @@ class MainActivity : AppCompatActivity() {
         getSupportActionBar()?.hide();
         setContentView(binding.root)
 
-        val navView: BottomNavigationView = binding.navView
+        val button = findViewById<Button>(R.id.button1)
 
-        val navController = findNavController(R.id.nav_host_fragment_activity_main)
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
-        val appBarConfiguration = AppBarConfiguration(
-            setOf(
-                R.id.navigation_home, R.id.navigation_courses, R.id.navigation_profile
-            )
-        )
-        setupActionBarWithNavController(navController, appBarConfiguration)
-        navView.setupWithNavController(navController)
+        button.setOnClickListener {
+            // Create an Intent object that specifies the activity to navigate to
+            val intent = Intent(this, InitialPage::class.java)
+
+            // Call the startActivity method and pass the Intent object as an argument
+            startActivity(intent)
+        }
+
+//        val navView: BottomNavigationView = binding.navView
+//
+//        val navController = findNavController(R.id.nav_host_fragment_activity_main)
+//        // Passing each menu ID as a set of Ids because each
+//        // menu should be considered as top level destinations.
+//        val appBarConfiguration = AppBarConfiguration(
+//            setOf(
+//                R.id.navigation_home, R.id.navigation_courses, R.id.navigation_profile
+//            )
+//        )
+//        setupActionBarWithNavController(navController, appBarConfiguration)
+//        navView.setupWithNavController(navController)
     }
 }
