+++ app/src/main/java/com/example/elearningmad/ui/EditProfile.kt
@@ -1,0 +1,12 @@
+package com.example.elearningmad.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.elearningmad.R
+
+class EditProfile : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_edit_profile)
+    }
+}
