+++ app/src/main/java/com/example/elearningmad/ui/InitialPage.kt
@@ -1,0 +1,37 @@
+package com.example.elearningmad.ui
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.findNavController
+import androidx.navigation.ui.AppBarConfiguration
+import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import com.example.elearningmad.R
+import com.example.elearningmad.databinding.ActivityInitialPageBinding
+import com.google.android.material.bottomnavigation.BottomNavigationView
+
+class InitialPage : AppCompatActivity() {
+
+    private lateinit var binding: ActivityInitialPageBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+//        setContentView(R.layout.activity_initial_page)
+
+        binding = ActivityInitialPageBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        val navView: BottomNavigationView = binding.navView
+
+        val navController = findNavController(R.id.nav_host_fragment_activity_main)
+        // Passing each menu ID as a set of Ids because each
+        // menu should be considered as top level destinations.
+        val appBarConfiguration = AppBarConfiguration(
+            setOf(
+                R.id.navigation_home, R.id.navigation_courses, R.id.navigation_profile
+            )
+        )
+        setupActionBarWithNavController(navController, appBarConfiguration)
+        navView.setupWithNavController(navController)
+    }
+}
