+++ app/src/main/res/layout/fragment_profile.xml
@@ -57,7 +57,7 @@
             android:text="Student number - IT21246296" />
 
     <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/button"
+        android:id="@+id/ProfileEditButton"
         android:layout_width="200dp"
         android:layout_height="50dp"
         android:layout_marginTop="10dp"
